major refactoring of browsing code to make it easier to add functionality

+++ brozzler/chrome.py
@@ -34,15 +34,23 @@ import tempfile
 class Chrome:
     logger = logging.getLogger(__module__ + '.' + __qualname__)
 
-    def __init__(
-            self, port, executable, proxy=None, ignore_cert_errors=False,
-            cookie_db=None):
+    def __init__(self, chrome_exe, port=9222, ignore_cert_errors=False):
+        '''
+        Initializes instance of this class.
+
+        Doesn't start the browser, start() does that.
+
+        Args:
+            chrome_exe: filesystem path to chrome/chromium executable
+            port: chrome debugging protocol port (default 9222)
+            ignore_cert_errors: configure chrome to accept all certs (default
+                False)
+        '''
         self.port = port
-        self.executable = executable
-        self.proxy = proxy
+        self.chrome_exe = chrome_exe
         self.ignore_cert_errors = ignore_cert_errors
-        self.cookie_db = cookie_db
         self._shutdown = threading.Event()
+        self.chrome_process = None
 
     def __enter__(self):
         '''
@@ -71,21 +79,19 @@ class Chrome:
 
         return default_port
 
-    def _init_cookie_db(self):
-        if self.cookie_db is not None:
-            cookie_dir = os.path.join(self._chrome_user_data_dir, 'Default')
-            cookie_location = os.path.join(cookie_dir, 'Cookies')
-            self.logger.debug(
-                    'cookie DB provided, writing to %s', cookie_location)
-            os.makedirs(cookie_dir, exist_ok=True)
+    def _init_cookie_db(self, cookie_db):
+        cookie_dir = os.path.join(self._chrome_user_data_dir, 'Default')
+        cookie_location = os.path.join(cookie_dir, 'Cookies')
+        self.logger.debug('cookie DB provided, writing to %s', cookie_location)
+        os.makedirs(cookie_dir, exist_ok=True)
 
-            try:
-                with open(cookie_location, 'wb') as cookie_file:
-                    cookie_file.write(self.cookie_db)
-            except OSError:
-                self.logger.error(
-                        'exception writing cookie file at %s',
-                        cookie_location, exc_info=True)
+        try:
+            with open(cookie_location, 'wb') as cookie_file:
+                cookie_file.write(cookie_db)
+        except OSError:
+            self.logger.error(
+                    'exception writing cookie file at %s',
+                    cookie_location, exc_info=True)
 
     def persist_and_read_cookie_db(self):
         cookie_location = os.path.join(
@@ -110,21 +116,32 @@ class Chrome:
                     cookie_location, exc_info=True)
         return cookie_db
 
-    def start(self):
+    def start(self, proxy=None, cookie_db=None):
         '''
-        Returns websocket url to chrome window with about:blank loaded.
+        Starts chrome/chromium process.
+
+        Args:
+            proxy: http proxy 'host:port' (default None)
+            cookie_db: raw bytes of chrome/chromium sqlite3 cookies database,
+                which, if supplied, will be written to
+                {chrome_user_data_dir}/Default/Cookies before running the
+                browser (default None)
+
+        Returns:
+            websocket url to chrome window with about:blank loaded
         '''
         # these can raise exceptions
         self._home_tmpdir = tempfile.TemporaryDirectory()
         self._chrome_user_data_dir = os.path.join(
             self._home_tmpdir.name, 'chrome-user-data')
-        self._init_cookie_db()
+        if cookie_db:
+            self._init_cookie_db(cookie_db)
 
         new_env = os.environ.copy()
         new_env['HOME'] = self._home_tmpdir.name
         self.port = self._find_available_port(self.port)
         chrome_args = [
-                self.executable, '--use-mock-keychain', # mac thing
+                self.chrome_exe, '--use-mock-keychain', # mac thing
                 '--user-data-dir=%s' % self._chrome_user_data_dir,
                 '--remote-debugging-port=%s' % self.port,
                 '--disable-web-sockets', '--disable-cache',
@@ -135,8 +152,8 @@ class Chrome:
                 '--disable-extensions', '--disable-save-password-bubble']
         if self.ignore_cert_errors:
             chrome_args.append('--ignore-certificate-errors')
-        if self.proxy:
-            chrome_args.append('--proxy-server=%s' % self.proxy)
+        if proxy:
+            chrome_args.append('--proxy-server=%s' % proxy)
         chrome_args.append('about:blank')
         self.logger.info(
                 'running: %s', repr(subprocess.list2cmdline(chrome_args)))
@@ -146,7 +163,8 @@ class Chrome:
                 stdout=subprocess.PIPE, stderr=subprocess.PIPE, bufsize=0)
         self._out_reader_thread = threading.Thread(
                 target=self._read_stderr_stdout,
-                name='ChromeOutReaderThread(pid=%s)' % self.chrome_process.pid)
+                name='ChromeOutReaderThread(pid=%s)' % self.chrome_process.pid,
+                daemon=True)
         self._out_reader_thread.start()
         self.logger.info('chrome running, pid %s' % self.chrome_process.pid)
 
@@ -171,6 +189,8 @@ class Chrome:
                             'got chrome window websocket debug url %s from %s',
                             url, json_url)
                     return url
+            except brozzler.ShutdownRequested:
+                raise
             except BaseException as e:
                 if int(time.time() - self._start) % 10 == 5:
                     self.logger.warn(
@@ -236,18 +256,18 @@ class Chrome:
     def stop(self):
         if not self.chrome_process or self._shutdown.is_set():
             return
+        self._shutdown.set()
 
         timeout_sec = 300
-        self._shutdown.set()
-        self.logger.info('terminating chrome pgid %s' % self.chrome_process.pid)
+        if self.chrome_process.poll() is None:
+            self.logger.info(
+                    'terminating chrome pgid %s', self.chrome_process.pid)
 
-        os.killpg(self.chrome_process.pid, signal.SIGTERM)
-        first_sigterm = time.time()
+            os.killpg(self.chrome_process.pid, signal.SIGTERM)
+        t0 = time.time()
 
         try:
-            while time.time() - first_sigterm < timeout_sec:
-                time.sleep(0.5)
-
+            while time.time() - t0 < timeout_sec:
                 status = self.chrome_process.poll()
                 if status is not None:
                     if status == 0:
@@ -264,11 +284,12 @@ class Chrome:
                     # around, but there's a chance I suppose that some other
                     # process could have started with the same pgid
                     return
+                time.sleep(0.5)
 
             self.logger.warn(
                     'chrome pid %s still alive %.1f seconds after sending '
                     'SIGTERM, sending SIGKILL', self.chrome_process.pid,
-                    time.time() - first_sigterm)
+                    time.time() - t0)
             os.killpg(self.chrome_process.pid, signal.SIGKILL)
             status = self.chrome_process.wait()
             self.logger.warn(
@@ -284,3 +305,4 @@ class Chrome:
         finally:
             self._out_reader_thread.join()
             self.chrome_process = None
+
